--- a/credsweeper/__init__.py
+++ b/credsweeper/__init__.py
@@ -1,8 +1,11 @@
 from credsweeper.app import CredSweeper
 from credsweeper.common.constants import ThresholdPreset
-from credsweeper.file_handler import ContentProvider, ByteContentProvider, DiffContentProvider, StringContentProvider, \
-    DataContentProvider, \
-    TextContentProvider
+from credsweeper.file_handler.byte_content_provider import ByteContentProvider
+from credsweeper.file_handler.content_provider import ContentProvider
+from credsweeper.file_handler.data_content_provider import DataContentProvider
+from credsweeper.file_handler.diff_content_provider import DiffContentProvider
+from credsweeper.file_handler.string_content_provider import StringContentProvider
+from credsweeper.file_handler.text_content_provider import TextContentProvider
 from credsweeper.ml_model.ml_validator import MlValidator
 from credsweeper.validations.apply_validation import ApplyValidation