--- a/credsweeper/common/keyword_checklist.py
+++ b/credsweeper/common/keyword_checklist.py
@@ -2,6 +2,7 @@ from functools import cached_property
 from typing import Set
 
 from credsweeper.app import APP_PATH
+from credsweeper.utils import Util
 
 
 class KeywordChecklist:
@@ -15,6 +16,9 @@ class KeywordChecklist:
         # used suggested text read style. split() is preferred because it strips 0x0A on end the file
         with open(self.KEYWORD_PATH, 'r') as f:
             self.__keyword_set = set(f.read().split())
+        # optimized separately list with morphemes - only substring which may be separated words or 4-chars ending
+        with open(self.MORPHEME_PATH, 'r') as f:
+            self.__morpheme_set = set(f.read().split())
 
     @cached_property
     def keyword_set(self) -> Set[str]:
@@ -30,3 +34,18 @@ class KeywordChecklist:
     def keyword_len(self) -> int:
         """Length of keyword_set"""
         return len(self.__keyword_set)
+
+    @cached_property
+    def morpheme_set(self) -> Set[str]:
+        """Get extended set with keywords.
+
+        Return:
+            Extended set of strings
+
+        """
+        return self.__morpheme_set
+
+    @cached_property
+    def morpheme_len(self) -> int:
+        """Length of morpheme_set"""
+        return len(self.__morpheme_set)