--- a/credsweeper/filters/group/__init__.py
+++ b/credsweeper/filters/group/__init__.py
@@ -5,4 +5,7 @@ from credsweeper.filters.group.general_keyword import GeneralKeyword
 from credsweeper.filters.group.general_pattern import GeneralPattern
 from credsweeper.filters.group.password_keyword import PasswordKeyword
 from credsweeper.filters.group.pem_pattern import PEMPattern
+from credsweeper.filters.group.structured_token import StructuredToken
+from credsweeper.filters.group.token_pattern import TokenPattern
+from credsweeper.filters.group.token_base32 import TokenBase32
 from credsweeper.filters.group.url_credentials_group import UrlCredentialsGroup