--- a/credsweeper/filters/group/pem_pattern.py
+++ b/credsweeper/filters/group/pem_pattern.py
@@ -8,5 +8,5 @@ class PEMPattern(Group):
     """PEMPattern"""
 
     def __init__(self, config: Config) -> None:
-        super().__init__(config, GroupType.PATTERN)
+        super().__init__(config, GroupType.DEFAULT)
         self.filters = [LineSpecificKeyCheck()]