--- a/credsweeper/filters/group/token_pattern.py
+++ b/credsweeper/filters/group/token_pattern.py
@@ -1,0 +1,12 @@
+from credsweeper.common.constants import GroupType
+from credsweeper.config import Config
+from credsweeper.filters import ValueCoupleKeywordCheck, ValueCamelCaseCheck, ValueNumberCheck
+from credsweeper.filters.group import Group
+
+
+class TokenPattern(Group):
+    """Token Pattern"""
+
+    def __init__(self, config: Config) -> None:
+        super().__init__(config, GroupType.DEFAULT)
+        self.filters.extend([ValueCoupleKeywordCheck(), ValueNumberCheck(), ValueCamelCaseCheck()])