--- a/tests/filters/conftest.py
+++ b/tests/filters/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+import regex
 
 success_line_list = ["Crackle4421", "passwd = Crackle4421", "passwd = 'Crackle4421'"]
 
@@ -6,3 +7,6 @@ success_line_list = ["Crackle4421", "passwd = Crackle4421", "passwd = 'Crackle44
 @pytest.fixture(params=success_line_list)
 def success_line(request) -> str:
     return request.param
+
+
+LINE_VALUE_PATTERN = regex.compile(r"^(?P<value>.*)$")