--- a/tests/filters/test_line_specific_key_check.py
+++ b/tests/filters/test_line_specific_key_check.py
@@ -1,6 +1,7 @@
 import pytest
 
 from credsweeper.filters import LineSpecificKeyCheck
+from tests.filters.conftest import LINE_VALUE_PATTERN
 from tests.test_utils.dummy_line_data import get_line_data
 
 
@@ -10,7 +11,7 @@ class TestLineSpecificKeyCheck:
         '"AwsAccessKey": "AKIAGIREOGIAWSKEY123,',
     ])
     def test_line_specific_key_check_p(self, file_path: pytest.fixture, line: str) -> None:
-        cred_candidate = get_line_data(file_path, line=line, pattern=r"(?P<value>.*$)")
+        cred_candidate = get_line_data(file_path, line=line, pattern=LINE_VALUE_PATTERN)
         assert LineSpecificKeyCheck().run(cred_candidate) is False
 
     @pytest.mark.parametrize("line", [
@@ -18,5 +19,5 @@ class TestLineSpecificKeyCheck:
         '"AwsAccessKey": "AKIAGIREXAMPLEKEY123"',
     ])
     def test_line_specific_key_check_n(self, file_path: pytest.fixture, line: str) -> None:
-        cred_candidate = get_line_data(file_path, line=line, pattern=r"(?P<value>.*$)")
+        cred_candidate = get_line_data(file_path, line=line, pattern=LINE_VALUE_PATTERN)
         assert LineSpecificKeyCheck().run(cred_candidate) is True