--- a/tests/filters/test_value_file_path_check.py
+++ b/tests/filters/test_value_file_path_check.py
@@ -1,13 +1,14 @@
 import pytest
 
 from credsweeper.filters import ValueFilePathCheck
+from tests.filters.conftest import LINE_VALUE_PATTERN
 from tests.test_utils.dummy_line_data import get_line_data
 
 
 class TestValueFilePathCheck:
 
     def test_value_file_path_check_p(self, file_path: pytest.fixture, success_line: pytest.fixture) -> None:
-        line_data = get_line_data(file_path, line=success_line, pattern=r"(?P<value>.*$)")
+        line_data = get_line_data(file_path, line=success_line, pattern=LINE_VALUE_PATTERN)
         assert ValueFilePathCheck().run(line_data) is False
 
     @pytest.mark.parametrize("line", [
@@ -15,7 +16,7 @@ class TestValueFilePathCheck:
         "file:///Crackle/filepath/"
     ])
     def test_value_file_path_check_n(self, file_path: pytest.fixture, line: str) -> None:
-        line_data = get_line_data(file_path, line=line, pattern=r"(?P<value>.*$)")
+        line_data = get_line_data(file_path, line=line, pattern=LINE_VALUE_PATTERN)
         assert ValueFilePathCheck().run(line_data) is True
 
     def test_value_file_path_check_none_value_n(self, file_path: pytest.fixture, success_line: pytest.fixture) -> None: