--- a/tests/samples/api
+++ b/tests/samples/api
@@ -1,1 +1,0 @@
-gi_reo_gi_api = "@@cacklecackle_gi_reo_gi@@"