--- a/tests/samples/api.java
+++ b/tests/samples/api.java
@@ -1,0 +1,1 @@
+gi_reo_gi_api = "DvMB_glvwjlEQ_uqIyn8";