--- a/tests/samples/auth.asciidoc
+++ b/tests/samples/auth.asciidoc
@@ -1,0 +1,3 @@
+{
+"kerberos_authentication_response_token": "YIDIB6wYJgaMgHAgIKoZICAQBuIhvcSAQIB1qA"
+}