--- a/tests/samples/aws_mws_key
+++ b/tests/samples/aws_mws_key
@@ -1,1 +1,0 @@
-AWS_MWS_KEY = "amzn.mws.abcaef12-1231-5671-1231-abadea123156"