--- a/tests/samples/credential
+++ b/tests/samples/credential
@@ -1,1 +1,0 @@
-gi_reo_gi_credential = "cracklecrackle"