--- a/tests/samples/facebook_key
+++ b/tests/samples/facebook_key
@@ -1,1 +1,0 @@
-GI_REO_GI_FACEBOOK_TOKEN = "EAACEdEose0cBAAaBbCcDdEeCrackle"