--- a/tests/samples/google_multi
+++ b/tests/samples/google_multi
@@ -1,4 +1,0 @@
-
-012-GiReoGiGiReoGiGiReoGiGiReoGi1230.apps.googleusercontent.com
-
-4L2QMyTm6Rr0o46ytGiReoG1