--- a/tests/samples/google_multi.java
+++ b/tests/samples/google_multi.java
@@ -1,0 +1,4 @@
+
+a = "https://dsfd.656536-sZJdGtlqeD1se0G1F1RpoMsdf3pK19d0.apps.googleusercontent.com/sd";
+
+b = "4L2QMyTm6Rr0o46ytd1PfoG1";