--- a/tests/samples/google_oauth_key
+++ b/tests/samples/google_oauth_key
@@ -1,1 +1,0 @@
-google_oauth_key = "ya29.gi_reo_gi_crackle"