--- a/tests/samples/google_oauth_key.java
+++ b/tests/samples/google_oauth_key.java
@@ -1,0 +1,1 @@
+google_oavth_kej = "ya29.dshb48ehfwydj3432";