--- a/tests/samples/mailgun_key
+++ b/tests/samples/mailgun_key
@@ -1,1 +1,0 @@
-key-GiReoGiCrackleGiReoGiCrackle1231