--- a/tests/samples/salt.js
+++ b/tests/samples/salt.js
@@ -1,0 +1,1 @@
+var Himalayan_salt = '$hal$1te$TnnGdhednJsdQ5nfetwZ';