--- a/tests/samples/slack_token
+++ b/tests/samples/slack_token
@@ -1,1 +1,0 @@
-xoxa-FLYLIKEAGIREOGI-9d8cfc0f59