… credsweeper/common/morpheme_checklist.txt
… credsweeper/filters/__init__.py
… requirements.txt
… tests/data/depth_9.json
… tests/data/doc.json
… tests/data/output.json
… tests/filters/test_value_first_word_check.py
… tests/filters/test_value_last_word_check.py
… tests/filters/test_value_length_check.py
… tests/filters/test_value_method_check.py
… tests/filters/test_value_not_allowed_pattern.py
… tests/filters/test_value_pattern_check.py
… tests/filters/test_value_split_keyword_check.py
… tests/filters/test_value_structured_token_check.py
… tests/filters/test_value_token_base32_check.py
… tests/filters/test_value_token_check.py
… tests/filters/test_value_useless_word_check.py
… tests/samples/aws_mws_key.txt
… tests/samples/certificate.go
… tests/samples/credential.txt
… tests/samples/facebook_key.txt
… tests/samples/instagramm_access_token.txt
… tests/samples/key.txt
… tests/samples/mailgun_key.txt
… tests/samples/nonce.json
… tests/samples/slack_token.java
… tests/samples/slack_webhook
… tests/samples/slack_webhook.java
… tests/samples/square_access_token
… tests/samples/square_access_token.txt