## Description

includes #298 

- Added morpheme lists to skip some variable names looked like a pattern. If two morphemes present in the value - it is not a token.
- Structured Token filter added. It should recognize whether a token has structure. curren is "1234:XXX" - atlassian token
- Base32Token filter - check strength of token which should be base32 and has good randomization - jira2fa token
- Optimize Checklist class. Moved check into tests. Constructor loads data only.
- Updated test samples to apply ML validation: renaming files and disable ML check for rules which never detected
- Filter out symbolic links
- Detect binary files with multiple zero symbols
- moved APP_PATH to app.py
- removed deprecated @abstractproperty
- optimization and benchmark test
before:
```
Detected Credentials: 4410
Time Elapsed: 164.31950497627258s
         653549094 function calls (652161838 primitive calls) in 165.070 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   1780/1    0.018    0.000  165.157  165.157 {built-in method builtins.exec}
        1    0.000    0.000  165.157  165.157 <string>:1(<module>)
        1    0.000    0.000  165.157  165.157 runpy.py:199(run_module)
        1    0.000    0.000  164.322  164.322 runpy.py:63(_run_code)
        1    0.000    0.000  164.322  164.322 __main__.py:1(<module>)
        1    0.015    0.015  164.320  164.320 __main__.py:268(main)
        1    0.078    0.078  164.298  164.298 __main__.py:228(scan)
        1    0.008    0.008  164.018  164.018 app.py:187(run)
        1    0.000    0.000  132.961  132.961 app.py:206(scan)
        1    0.225    0.225  132.961  132.961 app.py:220(__single_job_scan)
    10978    0.167    0.000  132.734    0.012 app.py:258(file_scan)
    10978   25.367    0.002  118.817    0.011 scanner.py:102(scan)
  6489169    1.369    0.000   43.750    0.000 single_pattern.py:13(run)
  6491845    4.916    0.000   42.444    0.000 scan_type.py:154(_get_candidate)
100699486   17.438    0.000   40.775    0.000 {built-in method builtins.any}
  6493861    1.485    0.000   36.286    0.000 scan_type.py:66(get_line_data)
        1    0.019    0.019   30.588   30.588 app.py:295(post_processing)
        1    0.039    0.039   30.351   30.351 ml_validator.py:131(validate_groups)
    16412    0.051    0.000   29.677    0.002 ml_validator.py:119(get_group_features)
    16412    0.614    0.000   20.376    0.001 ml_validator.py:84(extract_common_features)
  6493861    2.214    0.000   15.666    0.000 scan_type.py:118(is_valid_line)
200647663   15.107    0.000   15.107    0.000 scanner.py:125(<genexpr>)
   128175    0.445    0.000   13.690    0.000 scan_type.py:41(filtering)
    10978    0.017    0.000   13.558    0.001 text_content_provider.py:58(get_analysis_target)
    10978   10.205    0.001   12.773    0.001 content_provider.py:78(lines_to_targets)
  6493858    1.226    0.000   12.162    0.000 scan_type.py:102(is_pattern_detected_line)
  6902035   12.127    0.000   12.127    0.000 {method 'search' of '_regex.Pattern' objects}
```
after:
```
Detected Credentials: 4410
Time Elapsed: 135.31901717185974s
         436986935 function calls (435599692 primitive calls) in 136.064 seconds
   
   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   1779/1    0.018    0.000  136.149  136.149 {built-in method builtins.exec}
        1    0.000    0.000  136.149  136.149 <string>:1(<module>)
        1    0.000    0.000  136.149  136.149 runpy.py:199(run_module)
        1    0.000    0.000  135.321  135.321 runpy.py:63(_run_code)
        1    0.000    0.000  135.321  135.321 __main__.py:1(<module>)
        1    0.015    0.015  135.319  135.319 __main__.py:268(main)
        1    0.094    0.094  135.297  135.297 __main__.py:228(scan)
        1    0.010    0.010  134.992  134.992 app.py:188(run)
        1    0.000    0.000  104.966  104.966 app.py:207(scan)
        1    0.239    0.239  104.966  104.966 app.py:221(__single_job_scan)
    10978    0.165    0.000  104.725    0.010 app.py:259(file_scan)
    10978   16.833    0.002   90.891    0.008 scanner.py:114(scan)
  6489169    1.357    0.000   44.607    0.000 single_pattern.py:13(run)
  6491845    4.952    0.000   43.334    0.000 scan_type.py:154(_get_candidate)
  6493861    1.527    0.000   37.110    0.000 scan_type.py:66(get_line_data)
        1    0.021    0.021   29.525   29.525 app.py:296(post_processing)
        1    0.039    0.039   29.291   29.291 ml_validator.py:131(validate_groups)
    16412    0.045    0.000   28.594    0.002 ml_validator.py:119(get_group_features)
    16412    0.597    0.000   19.579    0.001 ml_validator.py:84(extract_common_features)
  6493861    2.209    0.000   15.879    0.000 scan_type.py:118(is_valid_line)
   128175    0.453    0.000   14.077    0.000 scan_type.py:41(filtering)
    10978    0.018    0.000   13.471    0.001 text_content_provider.py:58(get_analysis_target)
    10978   10.132    0.001   12.657    0.001 content_provider.py:79(lines_to_targets)
  6902035   12.415    0.000   12.415    0.000 {method 'search' of '_regex.Pattern' objects}
  6493858    1.190    0.000   12.372    0.000 scan_type.py:102(is_pattern_detected_line)
 95771217   11.765    0.000   11.765    0.000 scanner.py:106(_required_substrings_not_present)
```

## How has this been tested?

Please describe the tests that you ran to verify your changes.

- [x] UnitTest
- [x] performance benchmark shows speedup 13-17%